Changed MenuItem floats into doubles. Commented out database attribute as it hasnt been implemented. Created History attribute for Worker.

+++ SENG202-Team-5/src/main/java/seng202/group5/MenuItem.java
@@ -10,11 +10,11 @@ public class MenuItem {
 
     private String name;
     private Recipe recipe;
-    private float cost;
+    private double cost;
     private String ID;
 
 
-    public MenuItem(String tempName, Recipe tempRecipe, float tempCost, String tempID) {
+    public MenuItem(String tempName, Recipe tempRecipe, double tempCost, String tempID) {
         name = tempName;
         recipe = tempRecipe;
         cost = tempCost;
@@ -44,7 +44,7 @@ public class MenuItem {
     }
 
 
-    public float getCost() {
+    public double getCost() {
         return cost;
     }
